Add some test cases
git-svn-id: http://npoi.googlecode.com/svn/trunk@304 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ main/HSSF/Model/LinkTable.cs
@@ -164,9 +164,9 @@ namespace NPOI.HSSF.Model
          *
          * @see org.apache.poi.hssf.model.LinkTable#LinkTable(int, WorkbookRecordList)
          */
-            public ExternalBookBlock(short numberOfSheets)
+            public ExternalBookBlock(int numberOfSheets)
             {
-                _externalBookRecord = SupBookRecord.CreateInternalReferences(numberOfSheets);
+                _externalBookRecord = SupBookRecord.CreateInternalReferences((short)numberOfSheets);
                 _externalNameRecords = new ExternalNameRecord[0];
                 _crnBlocks = new CRNBlock[0];
             }
@@ -273,7 +273,7 @@ namespace NPOI.HSSF.Model
 
         }
 
-        public LinkTable(short numberOfSheets, WorkbookRecordList workbookRecordList)
+        public LinkTable(int numberOfSheets, WorkbookRecordList workbookRecordList)
         {
             _workbookRecordList = workbookRecordList;
             _definedNames = new List<NameRecord>();
